+++ src/test/java/com/kuber/starwarstest/TestConfiguration.java
@@ -1,0 +1,50 @@
+package com.kuber.starwarstest;
+
+import com.fasterxml.jackson.databind.DeserializationFeature;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.PropertyNamingStrategy;
+import org.apache.http.HttpHost;
+import org.apache.http.client.HttpClient;
+import org.apache.http.impl.client.HttpClients;
+import org.apache.http.impl.conn.DefaultProxyRoutePlanner;
+import org.springframework.boot.web.client.RestTemplateBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
+import org.springframework.http.converter.HttpMessageConverter;
+import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder;
+import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
+import org.springframework.web.client.RestTemplate;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Configuration
+public class TestConfiguration {
+
+    @Bean
+    public RestTemplate getRestTemplate() {
+        var builder = new Jackson2ObjectMapperBuilder();
+        builder.propertyNamingStrategy(PropertyNamingStrategy.SNAKE_CASE);
+        List<HttpMessageConverter<?>> converters = new ArrayList<>();
+        converters.add(new MappingJackson2HttpMessageConverter(builder.build()));
+
+        var objectMapper = new ObjectMapper();
+        objectMapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
+
+        RestTemplate restTemplate = new RestTemplateBuilder()
+                .messageConverters(new MappingJackson2HttpMessageConverter(objectMapper))
+                .messageConverters(converters)
+                .build();
+
+        HttpHost httpHost = new HttpHost(
+                System.getProperty("http.proxyHost"),
+                Integer.parseInt(System.getProperty("http.proxyPort"))
+        );
+        DefaultProxyRoutePlanner defaultProxyRoutePlanner = new DefaultProxyRoutePlanner(httpHost);
+        HttpClient httpClient = HttpClients.custom().setRoutePlanner(defaultProxyRoutePlanner).build();
+        restTemplate.setRequestFactory(new HttpComponentsClientHttpRequestFactory(httpClient));
+        return restTemplate;
+    }
+
+}