+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2021-04-09
+### Added
+- Added dependencies of the mock-server and TestContainer for integration test;
+- Create external API integration test - Starwars mock;
+
 ## [0.1.0] - 2021-03-14
 ### Added
 - Added some libraries such as Kotlin language, mockserver, junit 5 and netflix hystrix;